Move #with_associations method from AssociationProxy to QueryProxy
