+++ lib/neo4j.rb
@@ -55,6 +55,7 @@ require 'neo4j/active_node/query_methods'
 require 'neo4j/active_node/query/query_proxy_methods'
 require 'neo4j/active_node/query/query_proxy_enumerable'
 require 'neo4j/active_node/query/query_proxy_find_in_batches'
+require 'neo4j/active_node/query/query_proxy_eager_loading'
 require 'neo4j/active_node/query/query_proxy_link'
 require 'neo4j/active_node/labels'
 require 'neo4j/active_node/id_property'
